backend/requirements.txt
bcrypt==4.3.0
cryptography
fastapi
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
 